+++ ext/errors.go
@@ -2,5 +2,5 @@ package ext
 
 import "errors"
 
-var ErrExtAlreadyExists = errors.New("extension already existing")
+var ErrExtAlreadyExists = errors.New("extension already exists")
 var ErrExtNotFound = errors.New("extension not found")